GH-2: Refactored to add the ComputerPlayer(Player) class.
